Change test-nn.c to use trainXorNn
Had to refactor trainXorNn expose training patterns to test-nn and
added getEpochs and getError so they can be returned.

This isn't the best and eventually trainXorNn should probably have
the patterns as parameters and return a structure with the Error and
Epochs. But for now I don't know how to do that with wasm so this will
do for now.